+++ crates/app/src/commands/docker/prune.rs
@@ -340,7 +340,13 @@ pub async fn prune(session: MoonSession) -> AppResult {
     );
 
     for project_id in &manifest.focused_projects {
-        toolchains.extend(workspace_graph.get_project(project_id)?.toolchains.clone());
+        toolchains.extend(
+            workspace_graph
+                .get_project(project_id)?
+                .get_enabled_toolchains()
+                .into_iter()
+                .cloned(),
+        );
     }
 
     // Do this later so we only run once for each platform instead of per project
@@ -350,7 +356,9 @@ pub async fn prune(session: MoonSession) -> AppResult {
             continue;
         }
 
-        if let Ok(platform) = PlatformManager::read().get_by_toolchain(&toolchain_id) {
+        if let Ok(platform) = PlatformManager::write().get_by_toolchain_mut(&toolchain_id) {
+            platform.setup_toolchain().await?;
+
             match platform.get_type() {
                 PlatformType::Bun => {
                     prune_bun(