+++ crates/app/src/systems/analyze.rs
@@ -8,11 +8,13 @@ use moon_console::{Checkpoint, Console};
 use moon_deno_platform::DenoPlatform;
 use moon_env_var::GlobalEnvBag;
 use moon_node_platform::NodePlatform;
+use moon_pdk_api::SetupToolchainInput;
 use moon_platform::PlatformManager;
 use moon_python_platform::PythonPlatform;
 use moon_rust_platform::RustPlatform;
 use moon_system_platform::SystemPlatform;
 use moon_toolchain::is_using_global_toolchains;
+use moon_toolchain_plugin::ToolchainRegistry;
 use moon_vcs::BoxedVcs;
 use proto_core::{ProtoEnvironment, flow::install::ProtoInstallError, is_offline};
 use proto_installer::*;
@@ -242,7 +244,10 @@ pub async fn register_platforms(
 }
 
 #[instrument]
-pub async fn load_toolchain() -> AppResult {
+pub async fn load_toolchain(
+    toolchain_registry: &ToolchainRegistry,
+    toolchain_config: &ToolchainConfig,
+) -> AppResult {
     // This isn't an action but we should also support skipping here!
     if should_skip_action("MOON_SKIP_SETUP_TOOLCHAIN").is_some() {
         return Ok(None);
@@ -251,6 +256,18 @@ pub async fn load_toolchain() -> AppResult {
     for platform in PlatformManager::write().list_mut() {
         platform.setup_toolchain().await?;
     }
+
+    toolchain_registry
+        .setup_toolchain_all(|registry, toolchain| SetupToolchainInput {
+            configured_version: toolchain_config
+                .plugins
+                .get(toolchain.id.as_str())
+                .and_then(|plugin| plugin.version.clone()),
+            context: registry.create_context(),
+            toolchain_config: registry.create_config(&toolchain.id, toolchain_config),
+            version: None,
+        })
+        .await?;
 
     Ok(None)
 }